move espresso config to new package
