+++ system_tests/common_test.go
@@ -2477,6 +2477,7 @@ func populateMachineDir(t *testing.T, cr *github.ConsensusRelease) string {
 func createTestL1BlockChain(t *testing.T, l1info info, withClientWrapper bool) (info, *ethclient.Client, *eth.Ethereum, *node.Node, *ClientWrapper, daprovider.BlobReader) {
 	return createTestL1BlockChainWithL1StackConfig(t, l1info, testhelpers.CreateStackConfigForTest(t.TempDir()), withClientWrapper)
 }
+
 func TestEspressoConfigParsing(t *testing.T) {
 	inputSource := map[string]interface{}{
 		"sequencer": true,
